+++ selfdrive/car/hyundai/carcontroller.py
@@ -60,6 +60,7 @@ class CarController:
     self.send_lfa_mfa_lkas = CP.flags & HyundaiFlags.SEND_LFA.value
     self.last_button_frame = 0
     self.pcmCruiseButtonDelay = 0
+    self.dynamicJerk = 0
     self.jerkStartLimit = 1.0
     self.jerkUpperLowerLimit = 12.0       
     self.speedCameraHapticEndFrame = 0
@@ -119,6 +120,7 @@ class CarController:
 
     # *** common hyundai stuff ***
     if self.frame % 100 == 0:
+      self.dynamicJerk = int(Params().get("DynamicJerk", encoding="utf8"))
       self.jerkStartLimit = float(int(Params().get("JerkStartLimit", encoding="utf8"))) * 0.1
       self.jerkUpperLowerLimit = float(int(Params().get("JerkUpperLowerLimit", encoding="utf8")))
       self.hapticFeedbackWhenSpeedCamera = int(Params().get("HapticFeedbackWhenSpeedCamera", encoding="utf8"))
@@ -251,21 +253,19 @@ class CarController:
           jerk_u = self.jerkUpperLowerLimit
           jerk_l = self.jerkUpperLowerLimit
           self.jerk_count = 0
-        elif actuators.longControlState == LongCtrlState.stopping:
-          jerk_u = 0.5
-          jerk_l = self.jerkUpperLowerLimit
-          #self.jerk_count = 0
-        elif True:
-          jerk_u = interp(jerk, [-0.1, 0, 0.2], [0, 1.0, jerk_max])  #jerk_u가 0이 아니면, KONA_EV는 감속을 안함.
+        #elif actuators.longControlState == LongCtrlState.stopping:
+        #  jerk_u = 0.0 #0.5
+        #  jerk_l = self.jerkUpperLowerLimit
+        #  #self.jerk_count = 0
+        #elif True:
+        elif self.dynamicJerk == 1:
+          if actuators.longControlState == LongCtrlState.stopping:
+            jerk = -2.0
+          jerk_u = interp(jerk, [-0.1, 0, 0.2], [0.0, 1.0, jerk_max])  #jerk_u가 0이 아니면, KONA_EV는 감속을 안함. over감속:upper를 +로 하면? 230930
           jerk_l = interp(jerk, [-0.1, 0, 0.5], [jerk_max, 0.5, 0.5])  #jerk_l이 0.5가 아니면 가속도가 안올라감.
-        #elif abs(jerk) < 0.05:
-        #  jerk_u = jerk_l = jerk_max
-        #elif accel < 0: # or jerk < 0:
-        #  jerk_u = jerk_max if jerk > 0 else 0.5
-        #  jerk_l = jerk_max
-        #else:
-        #  jerk_u = jerk_max
-        #  jerk_l = jerk_max if jerk < 0 else 0.5
+        else:
+          jerk_u = jerk_l = jerk_max
+
         can_sends.extend(hyundaican.create_acc_commands_mix_scc(self.CP, self.packer, CC.enabled, accel, jerk_u, jerk_l, int(self.frame / 2),
                                                       hud_control, set_speed_in_units, stopping, CC, CS, self.softHoldMode))
         self.accel_last = accel