+++ selfdrive/car/hyundai/interface.py
@@ -315,7 +315,7 @@ class CarInterface(CarInterfaceBase):
 
     ret.stoppingControl = True
     ret.startingState = False # True # startAccel을 적용하는 startingState를 막음.
-    ret.vEgoStarting = 0.3
+    ret.vEgoStarting = 0.2
     ret.vEgoStopping = 0.3
     ret.startAccel = 2.0
     ret.stoppingDecelRate = 1.2 # brake_travel/s while trying to stop