+++ selfdrive/common/params.cc
@@ -317,6 +317,7 @@ std::unordered_map<std::string, uint32_t> keys = {
     { "TFollowGap2", PERSISTENT },
     { "TFollowGap3", PERSISTENT },
     { "TFollowGap4", PERSISTENT },
+    { "DynamicJerk", PERSISTENT },
     { "JerkStartLimit", PERSISTENT },
     { "JerkUpperLowerLimit", PERSISTENT },
     { "OPKRTimeZone", PERSISTENT },